10
8. Clicking the "Next" button should change the pointer to the
next city listed on your about page (i.e. a city you have lived in or want to visit). When the first of the series of
locations is selected, the "Previous" button should disappear, and when the last location is selected, the "Next" button
should disappear. Otherwise, both buttons should be visible.